ABC-Portal/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
reportlab
matplotlib
requests
psycopg2-binary
python-dotenv
